dom/workers/test/test_bug1104064.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1104064</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

var worker = new Worker("bug1104064_worker.js");
worker.onmessage = function() {
  ok(true, "setInterval has been called twice.");
  SimpleTest.finish();
}
worker.postMessage("go");

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
